@@ -31,7 +31,7 @@
     "format.check": "prettier packages projects scripts --check --config .prettierrc",
     "format": "prettier --write packages projects scripts --check --config .prettierrc",
     "format.scripts": "prettier --write ./scripts/src/**/*.{mts,ts,js,json}",
-    "lint": "ng lint",
+    "lint": "ng lint --quiet",
     "lint.fix": "ng lint --fix",
     "e2e": "npm run prestart && npm run e2e -w demo",
     "e2e.local": "npm run prestart &&npm run e2e.local -w demo",
@@ -94,18 +94,16 @@
     "@trivago/prettier-plugin-sort-imports": "^4.3.0",
     "@types/flexsearch": "^0.7.6",
     "@types/geojson": "^7946.0.10",
-    "@types/hammerjs": "^2.0.36",
     "@types/jasmine": "~5.1.0",
     "@types/lodash-es": "^4.17.0",
-    "@types/node": "^20.10.0",
+    "@types/node": "^22.13.0",
     "@types/tinycolor2": "^1.4.4",
     "angular-cli-ghpages": "^2.0.0",
-    "angular-eslint": "^18.4.0",
+    "angular-eslint": "^19.1.0",
     "conventional-changelog-cli": "^5.0.0",
     "copy-newer": "^2.1.2",
-    "cypress": "^13.16.0",
-    "del": "^8.0.0",
-    "eslint": "^9.15.0",
+    "cypress": "^14.0.0",
+    "eslint": "^9.21.0",
     "execa": "^9.5.0",
     "jasmine-core": "~5.1.0",
     "jasmine-spec-reporter": "~7.0.0",
@@ -117,8 +115,8 @@
     "lodash-es": "^4.17.21",
     "madge": "^8.0.0",
     "ng-packagr": "^19.1.2",
-    "prettier": "^3.4.0",
-    "sass": "^1.81.0",
+    "prettier": "^3.5.0",
+    "sass": "^1.85.0",
     "semantic-release": "^24.2.0",
     "tsx": "^4.19.0",
     "typescript": "~5.5.4",
