@@ -1,7 +1,7 @@
 import { Injectable, Optional } from '@angular/core';
 
 import { NgxIndexedDBService } from 'ngx-indexed-db';
-import { Observable, Subject, of } from 'rxjs';
+import { Observable, Subject } from 'rxjs';
 import { concatMap, first } from 'rxjs/operators';
 
 import { LayerDBData } from './layerDB.interface';
